requirements.txt
hvac
pyaml
pyOpenSSL
cryptography